requirements.txt
asgiref==3.9.1
certifi==2025.8.3
coverage==7.10.7
Django==5.2.5
django-cors-headers==4.7.0
django-impersonate==1.9.5
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
freezegun==1.5.5
gunicorn==23.0.0
packaging==25.0
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
sentry-sdk==2.35.2
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0
whitenoise==6.9.0
